+++ src/forms/App.js
@@ -4,8 +4,8 @@ Copyright © 2018 Roman Nep <neproman@gmail.com>
 */
 
 import ProxyPolyfill from 'proxy-polyfill/src/proxy';
-import { App } from 'kate-client';
-import { Layout, components } from 'kate-form-material-kit-react';
+import { App } from '../kate-client';
+import { Layout, components } from '../kate-form-material-kit-react';
 
 import makeItemForm from './Item';
 import makeListForm from './List';