+++ src/kate-client/reducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import { reducer } from 'kate-form';
+import { reducer } from '../kate-form';
 
 const rootReducer = combineReducers({
   'kate-form': reducer,